+++ changelog.md
@@ -11,3 +11,7 @@ All notable changes to this project will be documented in this file.
 *   Added repositories for all entities.
 *   Added Swagger configuration.
 *   Updated UserSleepLog to record sleep and wake up times.
+*   Added createUser, foodCategory, and foodType fields to Food entity.
+*   Added WeightRecord entity.
+*   Created FoodCategory and FoodType entities.
+*   Modified Food entity to use FoodCategory and FoodType entities.